# openapi-springboot.yaml
openapi: 3.0.3
info:
  title: SKAX Physical Risk Management API (Spring Boot)
  version: 1.0.0
  description: >
    사업장 기후 물리적 리스크 관리 시스템 백엔드 API

    프론트엔드와 직접 연결되며, AI Agent 분석은 내부적으로 FastAPI를 호출하여 제공합니다.

    주요 기능:
    - 사용자 인증 및 관리
    - 사업장 CRUD
    - AI 리스크 분석 결과 제공 (FastAPI 호출)
    - 리포트 생성 및 관리
    - 시뮬레이션
    - 대시보드
servers:
  - url: https://api.skax.co.kr
    description: Production
  - url: https://staging-api.skax.co.kr
    description: Staging
  - url: http://localhost:8080
    description: Local Development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: INVALID_REQUEST
        message:
          type: string
          example: 잘못된 요청입니다.
        timestamp:
          type: string
          format: date-time

    # ==========================================
    # Auth / User Schemas
    # ==========================================
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: user@skax.co.kr
        password:
          type: string
          format: password
          minLength: 8
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT Access Token
        refreshToken:
          type: string
          description: JWT Refresh Token
        user:
          $ref: '#/components/schemas/User'
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          description: 담당자 이름
          minLength: 2
        password:
          type: string
          format: password
          minLength: 8
        organization:
          type: string
          description: 소속 조직
        verificationCode:
          type: string
          description: 이메일 인증번호
    RefreshTokenRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
    PasswordResetRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    PasswordResetConfirmRequest:
      type: object
      required: [token, newPassword]
      properties:
        token:
          type: string
          description: 이메일로 발송된 재설정 토큰
        newPassword:
          type: string
          format: password
          minLength: 8
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        organization:
          type: string
        language:
          type: string
          example: ko
          enum: [ko, en]
        role:
          type: string
          enum: [USER, ADMIN]
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
...
    Site:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: 대덕 데이터센터
        address:
          type: string
          example: 대전광역시 유성구 대덕대로 512
        city:
          type: string
          example: 대전광역시
        latitude:
          type: number
          format: double
          example: 36.368
        longitude:
          type: number
          format: double
          example: 127.364
        industry:
          type: string
          example: DC(Data Center)
          description: 업종
        buildingAge:
          type: integer
          description: 건물 연령 (년)
          example: 15
        buildingType:
          type: string
          description: 건물 유형
          example: 철근 콘크리트
        seismicDesign:
          type: boolean
          description: 내진설계 여부
          default: false
        floorArea:
          type: number
          format: double
          description: 연면적 (m²)
          example: 5000.0
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
          example: 50000000000
        employeeCount:
          type: integer
          description: 직원 수
          example: 100
        mainHazard:
          type: string
          example: 폭염
          description: 주요 리스크
        riskScore:
          type: integer
          minimum: 0
          maximum: 100
          example: 70
          description: 종합 리스크 점수
        riskLevel:
          type: string
          enum: [낮음, 보통, 높음, 심각]
          example: 높음
        lastAnalyzedAt:
          type: string
          format: date-time
          description: 마지막 분석 일시
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateSiteRequest:
      type: object
      required: [name, address, industry, latitude, longitude]
      properties:
        name:
          type: string
          minLength: 2
        address:
          type: string
          minLength: 5
        industry:
          type: string
          example: DC(Data Center)
        latitude:
          type: number
          format: double
          description: 위도 (필수)
        longitude:
          type: number
          format: double
          description: 경도 (필수)
        buildingAge:
          type: integer
          description: 건물 연령 (년)
          example: 15
        buildingType:
          type: string
          description: 건물 유형
          example: 철근 콘크리트
        seismicDesign:
          type: boolean
          description: 내진설계 여부
          default: false
        floorArea:
          type: number
          format: double
          description: 연면적 (m²)
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
        employeeCount:
          type: integer
          description: 직원 수
        description:
          type: string
          description: 사업장 설명
    UpdateSiteRequest:
      type: object
      properties:
        name:
          type: string
        address:
          type: string
        industry:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        buildingAge:
          type: integer
          description: 건물 연령 (년)
        buildingType:
          type: string
          description: 건물 유형
        seismicDesign:
          type: boolean
          description: 내진설계 여부
        floorArea:
          type: number
          format: double
          description: 연면적 (m²)
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
        employeeCount:
          type: integer
          description: 직원 수
        description:
          type: string
    SiteListResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Site'
        totalElements:
          type: integer
        totalPages:
          type: integer
        currentPage:
          type: integer
        pageSize:
          type: integer

    # ==========================================
    # Dashboard Schemas
    # ==========================================
    DashboardSummary:
      type: object
      properties:
        totalSites:
          type: integer
          example: 12
        safeSites:
          type: integer
          example: 5
          description: 리스크 낮음
        riskySites:
          type: integer
          example: 7
          description: 리스크 높음 이상
        highestRiskSite:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            riskScore:
              type: integer
        highestRiskHazard:
          type: string
          example: 폭염
        recentAnalyses:
          type: array
          items:
            type: object
            properties:
              siteId:
                type: string
              siteName:
                type: string
              analyzedAt:
                type: string
                format: date-time
              riskScore:
                type: integer

    # ==========================================
    # Analysis Schemas (FastAPI 연동)
    # ==========================================
    PhysicalRiskBarItem:
      type: object
      properties:
        hazardType:
          type: string
          example: 홍수 빈도
        score:
          type: integer
          description: 물리적 리스크 종합 점수 (0-100)
        hazardScore:
          type: number
          format: float
          description: Hazard 점수 (0.0-1.0)
        exposureScore:
          type: number
          format: float
          description: Exposure 점수 (0.0-1.0)
        vulnerabilityScore:
          type: number
          format: float
          description: Vulnerability 점수 (0.0-1.0)
        riskCalculationMethod:
          type: string
          example: H×E×V
...
    PhysicalRiskScoreResponse:
      type: object
      description: Hazard별 물리적 리스크 점수
      properties:
        siteId:
          type: string
          format: uuid
        analyzedAt:
          type: string
          format: date-time
        hazardType:
          type: string
          example: 폭염
          description: 조회한 위험요인 (전체인 경우 null)
        physicalRiskScores:
          type: array
          description: Hazard별 물리적 리스크 점수
          items:
            $ref: '#/components/schemas/PhysicalRiskBarItem'
    PastEventSummary:
      type: object
      properties:
        totalEvents:
          type: integer
          description: 총 이벤트 수
        severeEvents:
          type: integer
          description: 심각한 이벤트 수
        avgWarningDays:
          type: number
          format: float
          description: 평균 경보 일수
        avgSevereDays:
          type: number
          format: float
          description: 평균 심각 일수
        mostFrequentHazard:
          type: string
          description: 가장 빈번한 위험요인
        yearRange:
          type: object
          properties:
            start:
              type: integer
            end:
              type: integer
...
    PastEventsResponse:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        summary:
          $ref: '#/components/schemas/PastEventSummary'
        events:
          type: array
          items:
            $ref: '#/components/schemas/PastEventItem'
...
    FinancialImpactItem:
      type: object
      properties:
        hazardType:
          type: string
          example: 폭염
          description: 위험요인 유형
        aal:
          type: number
          format: float
          description: 연평균 자산 손실률 (AAL, 0.0-1.0)
          example: 0.023
        annualExpectedLoss:
          type: number
          format: double
          description: 연평균 재무 손실액 (KRW) = AAL × 자산가치
          example: 230000000
    FinancialImpactResponse:
      type: object
      description: Hazard별 재무 영향 (AAL)
      properties:
        siteId:
          type: string
          format: uuid
        analyzedAt:
          type: string
          format: date-time
        hazardType:
          type: string
          example: 폭염
          description: 조회한 위험요인 (전체인 경우 null)
        assetValue:
          type: number
          format: double
          description: 자산 가치 (KRW)
          example: 10000000000
        financialImpacts:
          type: array
          description: Hazard별 재무 영향
          items:
            $ref: '#/components/schemas/FinancialImpactItem'
...
    VulnerabilityResponse:
      type: object
      properties:
        siteId:
          type: string
          format: uuid
        totalScore:
          type: integer
          minimum: 0
          maximum: 100
          description: 종합 취약성 점수
        overallLevel:
          type: string
          enum: [낮음, 보통, 높음, 심각]
        siteInfo:
          type: object
          description: 사업장 정보
          properties:
            buildingAge:
              type: integer
            buildingType:
              type: string
            seismicDesign:
              type: boolean
            floorArea:
              type: number
        details:
          type: array
          description: Hazard별 취약성 상세
          items:
            $ref: '#/components/schemas/VulnerabilityItem'
        priorityActions:
          type: array
          description: 우선 조치 사항
          items:
            type: object
            properties:
              action:
                type: string
              priority:
                type: string
                enum: [low, medium, high, critical]
              estimatedCost:
                type: number
    AnalysisTotalResponse:
      type: object
      description: 특정 Hazard 기준 통합 분석 결과
      properties:
        siteId:
          type: string
          format: uuid
        hazardType:
          type: string
          example: 폭염
        overview:
          $ref: '#/components/schemas/AnalysisOverviewResponse'
        pastEvents:
          $ref: '#/components/schemas/PastEventsResponse'
        sspProjection:
          $ref: '#/components/schemas/SSPProjectionResponse'
        financialImpact:
          $ref: '#/components/schemas/FinancialImpactResponse'
        vulnerability:
          $ref: '#/components/schemas/VulnerabilityResponse'
        executiveSummary:
          type: string
          description: AI가 생성한 경영진 요약 리포트
        generatedAt:
          type: string
          format: date-time
    StartAnalysisRequest:
      type: object
      properties:
        hazardTypes:
          type: array
          description: 분석할 리스크 유형 (미지정시 전체)
          items:
            type: string
          example: [폭염, 태풍, 홍수]
        priority:
          type: string
          enum: [low, normal, high]
          default: normal
    AnalysisJobStatus:
      type: object
      properties:
        jobId:
          type: string
        siteId:
          type: string
        status:
          type: string
          enum: [queued, running, completed, failed]
        progress:
          type: integer
          minimum: 0
          maximum: 100
        currentNode:
          type: string
          example: physical_risk_score
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        estimatedCompletionTime:
          type: string
          format: date-time
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string

    # ==========================================
    # Report Schemas
    # ==========================================
...
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: 로그인
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 로그인 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/register:
    post:
      tags: [Auth]
      summary: 회원가입
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: 회원가입 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 이미 존재하는 이메일
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Access Token 갱신
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: 토큰 갱신 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  accessToken:
                    type: string
                  refreshToken:
                    type: string
  /api/auth/logout:
    post:
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      summary: 로그아웃
      responses:
        '204':
          description: 로그아웃 성공
  /api/auth/password/reset-request:
    post:
      tags: [Auth]
      summary: 비밀번호 재설정 메일 요청
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetRequest'
      responses:
        '200':
          description: 메일 전송 완료
  /api/auth/password/reset-confirm:
    post:
      tags: [Auth]
      summary: 비밀번호 재설정 완료
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PasswordResetConfirmRequest'
      responses:
        '200':
          description: 비밀번호 재설정 성공

  # ==========================================
  # Users Endpoints
  # ==========================================
...
  /api/sites:
    get:
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      summary: 사업장 목록 조회
      parameters:
        - in: query
          name: search
          schema:
            type: string
          description: 이름/주소 검색
        - in: query
          name: riskLevel
          schema:
            type: string
            enum: [낮음, 보통, 높음, 심각]
          description: 리스크 레벨 필터
        - in: query
          name: industry
          schema:
            type: string
          description: 업종 필터
        - in: query
          name: page
          schema:
            type: integer
            default: 0
        - in: query
          name: size
          schema:
            type: integer
            default: 20
        - in: query
          name: sort
          schema:
            type: string
            default: createdAt,desc
      responses:
        '200':
          description: 사업장 목록
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SiteListResponse'

    post:
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      summary: 사업장 등록
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSiteRequest'
      responses:
        '201':
          description: 생성된 사업장
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
  /api/sites/{siteId}:
    get:
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      summary: 사업장 상세 조회
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 사업장 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'
        '404':
          description: 사업장을 찾을 수 없음

    patch:
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      summary: 사업장 정보 수정
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSiteRequest'
      responses:
        '200':
          description: 수정된 사업장
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Site'

    delete:
      tags: [Sites]
      security: [{ bearerAuth: [] }]
      summary: 사업장 삭제
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 삭제 완료

  # ==========================================
  # Dashboard Endpoints
  # ==========================================
  /api/dashboard/summary:
    get:
      tags: [Dashboard]
      security: [{ bearerAuth: [] }]
      summary: 대시보드 요약 정보
      responses:
        '200':
          description: 요약 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardSummary'

  # ==========================================
  # Analysis Endpoints (FastAPI 연동)
  # ==========================================
  /api/sites/{siteId}/analysis/start:
    post:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: AI 리스크 분석 시작
      description: >
        FastAPI AI Agent를 호출하여 사업장에 대한 리스크 분석을 시작합니다.
        비동기로 실행되며, jobId를 통해 진행 상태를 추적할 수 있습니다.
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartAnalysisRequest'
      responses:
        '202':
          description: 분석 작업 시작됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobStatus'
        '404':
          description: 사업장을 찾을 수 없음
        '409':
          description: 이미 실행 중인 작업이 있음
  /api/sites/{siteId}/analysis/status/{jobId}:
    get:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: 분석 작업 상태 조회
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
        - in: path
          name: jobId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 작업 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisJobStatus'
...
  /api/sites/{siteId}/analysis/physical-risk-scores:
    get:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: 물리적 리스크 점수 조회
      description: Hazard별 물리적 리스크 점수를 조회합니다.
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
            format: uuid
          description: 사업장 ID
        - in: query
          name: hazardType
          schema:
            type: string
            enum: [폭염, 한파, 태풍, 홍수, 가뭄, 산불, 해안침수, 도시침수, 물부족]
          description: 특정 Hazard만 조회 (미지정시 전체)
      responses:
        '200':
          description: 물리적 리스크 점수
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PhysicalRiskScoreResponse'
        '404':
          description: 분석 결과를 찾을 수 없음
  /api/sites/{siteId}/analysis/past-events:
    get:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: 과거 재난 이력
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 과거 재난
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PastEventsResponse'
...
  /api/sites/{siteId}/analysis/financial-impacts:
    get:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: 연평균 자산 손실률 조회
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
        - in: query
          name: hazardType
          schema:
            type: string
      responses:
        '200':
          description: 재무 영향
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FinancialImpactResponse'
  /api/sites/{siteId}/analysis/vulnerability:
    get:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: 취약성 분석
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 취약성 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VulnerabilityResponse'
  /api/sites/{siteId}/analysis/total:
    get:
      tags: [Analysis]
      security: [{ bearerAuth: [] }]
      summary: 특정 Hazard 기준 통합 분석 결과
      parameters:
        - in: path
          name: siteId
          required: true
          schema:
            type: string
        - in: query
          name: hazardType
          required: true
          schema:
            type: string
        - in: query
          name: timeScale
          schema:
            type: string
      responses:
        '200':
          description: 통합 분석 결과
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisTotalResponse'

  # ==========================================
  # Reports Endpoints
  # ==========================================
...